requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo
python-dotenv
scikit-learn==1.2.2
pandas
numpy